Fix cost accounting, GGUF resolution and teardown safety found in liv… (#7)
* Fix cost accounting, GGUF resolution and teardown safety found in live runs

- Resolve GGUF weights in the daemon after the quantisation default, so up, offers, TUI and MCP all get Q4_K_M, not BF16
- Let MCP and `larri tui --runtime llamacpp` use llama.cpp; they failed resolving an empty model ref
- Accept `owner/repo/path/file.gguf` refs: size by repository, measure, check against the listing, record the file's quantisation
- Serve a named GGUF file as `qwen3-8b-q8_0` rather than the file name or its first part
- Size a quantisation published twice as one download, so advice no longer offers a larger file as a saving
- Quote Vast offers for the disk actually rented (allocated_storage), so the quote matches the bill
- Read Vast storage_cost as $/GB/month, not $/hr; it added $0.20/hr to every rig
- Price RunPod container disk and volume beside the GPU rate in quotes and instance rates
- Journal the billed rate, count storage once, and mark entries with the rules they were written under
- Read old journal entries by their original rules, at the billed rate the rig file kept where available
- Write each rig's termination to the journal entry that ends it; the field existed but was never set
- Resolve failed creates: close never-sent ones, tear down labelled instances, keep unknown outcomes FAILED
- Stop cost only on recorded nothing_created evidence, never on a blank instance id
- Add `larri down --nothing-created "<how you checked>" <rig>` to correct records LARRI cannot settle
- Act on a rig only through the provider holding it; down, resume and MCP used the configured default
- Keep a destroyed rig destroyed: the store refuses revival and cost replay ignores entries after the end
- End supervision when another process destroys the rig, or when probes fail and the provider no longer has it
- Offer only RunPod GPU types its create schema accepts, which drops the PRO 4500 Server Edition
- Report llama.cpp download percentage while weights arrive
- Measure download progress in bytes written, not blocks the filesystem reserved
- Follow the symlinked Hugging Face cache on RunPod so vLLM progress no longer sits at 0%
- Remove unused sizing.IsGGUF and move its cases into the runtime-selection test
- Add a regression test that the SSH host-key scan never completes a handshake
- Update the requirements spec, PROJECT_STATE, design doc, CLAUDE.md, README and site pages

* Fixing the incorrect conversion of unsigned 64-bit int security issue by github

* Address PR #7 review: teardown races, incomplete GGUF sets, storage replay

- Keep a rig billable when `larri down` cannot ask the provider whether anything carries its label
- Refuse GGUF sets with a missing part, named or picked from the listing, before renting
- Accrue the billed rate while running and the storage rate only while stopped
- Choose the MCP runtime from the model, as the CLI does
- Adopt the first teardown's record when a stale second teardown arrives, at start or at write
- Fail closed on an unreadable rig file for every move except teardown
- List GGUF files at the revision that will be downloaded
- Date the operator's correction when the rig had no recorded ending

* Address second PR #7 review: orphan providers, snapshot locking, resume runtime

- Confirm an orphan is at the provider before destroying it; require the provider in larri_orphan_destroy
- Sweep orphans at every provider that opens, in the CLI and MCP, reporting any not checked
- Guard every rig snapshot write against reviving a destroyed rig, under a per-rig flock
- Stop a bring-up whose snapshot write finds the rig ended elsewhere
- Resume a rig with the runtime it was brought up with, not always vLLM
- Record a dated instance-gone ending when adopt finds the instance absent
- Refresh the instance from the provider before adopt journals a stopped rig
- Bound and cancel llama.cpp progress probes so a wedged du cannot hold Bootstrap
- Test the host-key scan on its own connection: it must never reach authentication
- Require a quantisation for a named GGUF file whose name carries none
- Match multi-part GGUF names case-insensitively, keeping the listed extension

+++ internal/runtime/fake/fake.go
@@ -112,8 +112,8 @@ func (r *Runtime) Bootstrap(ctx context.Context, _ runtime.Session, spec core.Mo
 		send(runtime.Progress{
 			Phase:      runtime.PhaseWeightsDownload,
 			Percent:    pct,
-			BytesDone:  int64(float64(r.downloadBytes()) * pct / 100),
-			BytesTotal: r.downloadBytes(),
+			BytesDone:  uint64(float64(r.downloadBytes()) * pct / 100),
+			BytesTotal: uint64(r.downloadBytes()),
 		})
 	}
 	return nil
